fix(tests) merged two test files to one
- merged courseroutes.spec.js and courseroutes.test.js
- updated the fakeCourse and fakeCourses model to slighly reflect the actual model more
- updated the tests to reflect the updated model
